// ecomts-rtk/src/model/types.ts
export type Product = {
    id: number,
    title: string,
    description : string,
    images: string[],
    price: number,
    thumbnail?:string,
    reviews?:string[]
}

export type CartItem = Product & {
    qty: number,
    amount: number
}


export interface User {
  id: number;
  name: string;
  username: string;
  email: string;
  address: Address;
  phone: string;
  website: string;
  company: Company;
}

interface Company {
  name: string;
  catchPhrase: string;
  bs: string;
}

interface Address {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
  geo: Geo;
}

interface Geo {
  lat: string;
  lng: string;
}